fix: fixes circular dependency
